Scene graph refactor and null client patch (#635)
* Fix null client patch build.
* Extend the runtime.

Add a templatized Handle for handle-based storage.
Split up prefab/model/axisalignedbox from scene graph
namespace the scene graph parts.

* Architectural changes:

* Handlers/Services now have a separate _interface libraries that users
of given service can be linked with, to provide access to all message classes given service supports.

* introduced 2 new cmake macros to help with adding new handlers:
** `add_messages_library` which scans current directory for headers with message macros
and creates _interface library
** `add_handler` creates a handler library from provided source files, and marks it to link with a correspondng _interface library
* NetStructures library was merged with GameData

Runtime data and scene graph changes:

* Scene graph now learned to properly store most of the information provided by the loaded files.
* Many parts of scene graph use handle based storage.
* Runtime data now also knows how to handle models/textures

MapViewer changes:
* updated to latest lutefisk3d
* no longer uses a custom scene graph loader.
* makes a better use of cached graph nodes.

Misc changes:
* Common needs ACE as a full blown dependency.
* All _interface classes in turn, depend on Common
* Visual studio project view has all targets in proper folders
* Fix output directory paths
* SEGSEventFactory learned how to pack/unpack events to a std::stream.

+++ Projects/CoX/Common/CRUDP_Protocol/CRUD_Events.h
@@ -48,7 +48,7 @@ public:
 class CRUDLink_Event : public Event
 {
 public:
-    CRUDLink_Event(size_t evtype, EventSrc *ev_src = nullptr) : Event(evtype, ev_src) //,LINK
+    CRUDLink_Event(uint32_t evtype, EventSrc *ev_src = nullptr) : Event(evtype, ev_src) //,LINK
     {
     }
     virtual void serializeto(BitStream &) const = 0;
